refactor(hooks): separate
